# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: National LMIS Ledger System
  description: Decoupled Modular Monolith for Supply Chain Accounting
  version: 1.0.0
paths:
...
  /api/ledger/commands:
    post:
      tags:
      - Ledger Core
      summary: Submit Ledger Command
      description: 'Internal API Gateway for Ledger operations.

        The Adapter (acting as ''ledger_system'') submits normalized commands here.

        This routes to Area B (Idempotency), Area E (Approval), or Area C/D directly.'
      operationId: submit_ledger_command_api_ledger_commands_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LedgerCommand'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/ledger/balances:
    get:
      tags:
      - Ledger Core
      summary: Get Stock Balances
      description: 'CQRS Read API: Fetches current pre-calculated stock balances.

        Automatically filters results based on the JWT `allowed_nodes` claims.

        Optionally filter to a specific valid node_id.'
      operationId: get_stock_balances_api_ledger_balances_get
      security:
      - HTTPBearer: []
      parameters:
      - name: node_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Node Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockBalanceResponse'
                title: Response Get Stock Balances Api Ledger Balances Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/ledger/history/{node_id}/{item_id}:
    get:
      tags:
      - Ledger Core
      summary: Get Inventory History
      description: 'CQRS Read API: Fetches the immutable event log explaining how
        a balance arrived at its current state.

        Strictly asserts the actor has access to the requested `node_id`.'
      operationId: get_inventory_history_api_ledger_history__node_id___item_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: node_id
        in: path
        required: true
        schema:
          type: string
          title: Node Id
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          title: Item Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LedgerHistoryResponse'
                title: Response Get Inventory History Api Ledger History  Node Id   Item
                  Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/ledger/transfers:
    get:
      tags:
      - Ledger Core
      summary: List Transfers
      description: 'CQRS Read API: Fetches pending incoming/outgoing transfers.

        Automatically filters results based on the JWT `allowed_nodes` claims.'
      operationId: list_transfers_api_ledger_transfers_get
      security:
      - HTTPBearer: []
      parameters:
      - name: node_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Node Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InTransitTransferResponse'
                title: Response List Transfers Api Ledger Transfers Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/ledger/transfers/{transfer_id}/receive:
    post:
      tags:
      - Ledger Core
      summary: Receive Transfer
      description: UI Endpoint to mark a dispatched transfer as received.
      operationId: receive_transfer_api_ledger_transfers__transfer_id__receive_post
      security:
      - HTTPBearer: []
      parameters:
      - name: transfer_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Transfer Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiveTransferRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/ledger/gatekeeper/staged:
    get:
      tags:
      - Ledger Gatekeeper
      summary: List Staged Commands
      description: List transactions awaiting approval for the actor's allowed nodes.
      operationId: list_staged_commands_api_ledger_gatekeeper_staged_get
      security:
      - HTTPBearer: []
      parameters:
      - name: node_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Node Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/ledger/gatekeeper/{staged_id}/resolve:
    post:
      tags:
      - Ledger Gatekeeper
      summary: Resolve Staged Command
      description: UI Endpoint for Supervisors to Approve or Reject transactions.
      operationId: resolve_staged_command_api_ledger_gatekeeper__staged_id__resolve_post
      security:
      - HTTPBearer: []
      parameters:
      - name: staged_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Staged Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalActionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApprovalActionRequest:
      properties:
        action:
          type: string
          title: Action
          description: APPROVE or REJECT
        comment:
          anyOf:
          - type: string
          - type: 'null'
          title: Comment
          description: Optional notes for the decision.
      type: object
      required:
      - action
      title: ApprovalActionRequest
      description: Strict schema for the UI to submit an approval action
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InTransitStatus:
      type: string
      enum:
      - OPEN
      - PARTIAL
      - COMPLETED
      - STALE_AUTO_CLOSED
      - FAILED_AUTO_CLOSE
      - LOST
      title: InTransitStatus
    InTransitTransferResponse:
      properties:
        transfer_id:
          type: string
          format: uuid
          title: Transfer Id
        source_node_id:
          type: string
          title: Source Node Id
        dest_node_id:
          type: string
          title: Dest Node Id
        item_id:
          type: string
          title: Item Id
        qty_shipped:
          type: integer
          title: Qty Shipped
        qty_received:
          type: integer
          title: Qty Received
        status:
          $ref: '#/components/schemas/InTransitStatus'
        dispatched_at:
          type: string
          format: date-time
          title: Dispatched At
        auto_close_after:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Auto Close After
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - transfer_id
      - source_node_id
      - dest_node_id
      - item_id
      - qty_shipped
      - qty_received
      - status
      - dispatched_at
      - created_at
      - updated_at
      title: InTransitTransferResponse
      description: Data Transfer Object representing an active or completed In-Transit
        shipment.
    LedgerCommand:
      properties:
        source_event_id:
          type: string
          title: Source Event Id
          description: The unique ID from the source (e.g., ODK InstanceID).
        version_timestamp:
          type: integer
          title: Version Timestamp
          description: Timestamp acting as version for edits/reversals.
        transaction_type:
          $ref: '#/components/schemas/TransactionType'
        node_id:
          type: string
          title: Node Id
          description: Internal ID of the facility or MU
        item_id:
          type: string
          title: Item Id
          description: Internal ID of the commodity (Shared Kernel)
        quantity:
          type: integer
          title: Quantity
          description: The delta. Always in Base Units.
        transfer_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Transfer Id
          description: Links a receipt to a specific dispatch.
        batch_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Batch Id
          description: For tracking specific lots/batches.
        expiry_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expiry Date
          description: Essential for FEFO logic.
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
          description: The 'Business Time' (when it happened in the field).
        metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metadata
          description: 'Contextual info (e.g., ''Reason: Damaged in transit'').'
      type: object
      required:
      - source_event_id
      - version_timestamp
      - transaction_type
      - node_id
      - item_id
      - quantity
      - occurred_at
      title: LedgerCommand
    LedgerHistoryResponse:
      properties:
        source_event_id:
          type: string
          title: Source Event Id
        transaction_type:
          type: string
          title: Transaction Type
        node_id:
          type: string
          title: Node Id
        item_id:
          type: string
          title: Item Id
        quantity:
          type: integer
          title: Quantity
        running_balance:
          type: integer
          title: Running Balance
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
      type: object
      required:
      - source_event_id
      - transaction_type
      - node_id
      - item_id
      - quantity
      - running_balance
      - occurred_at
      title: LedgerHistoryResponse
      description: View model for the immutable audit log underlying the balances
...
    ReceiveTransferRequest:
      properties:
        qty_received:
          type: integer
          title: Qty Received
        node_id:
          type: string
          title: Node Id
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
        source_event_id:
          type: string
          title: Source Event Id
      type: object
      required:
      - qty_received
      - node_id
      - occurred_at
      - source_event_id
      title: ReceiveTransferRequest
      description: Payload for completing an In-Transit receipt from the UI.
    StockBalanceResponse:
      properties:
        node_id:
          type: string
          title: Node Id
        item_id:
          type: string
          title: Item Id
        quantity:
          type: integer
          title: Quantity
        last_updated:
          type: string
          format: date-time
          title: Last Updated
      type: object
      required:
      - node_id
      - item_id
      - quantity
      - last_updated
      title: StockBalanceResponse
      description: View model for the current pre-calculated stock projection
    TransactionType:
      type: string
      enum:
      - RECEIPT
      - ISSUE
      - TRANSFER
      - ADJUSTMENT
      - STOCK_COUNT
      - REVERSAL
      - LOSS_IN_TRANSIT
      title: TransactionType
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer